+++ src/Service/Lambda/src/Exception/S3FilesMountTimeoutException.php
@@ -1,0 +1,32 @@
+<?php
+
+namespace AsyncAws\Lambda\Exception;
+
+use AsyncAws\Core\Exception\Http\ClientException;
+use Symfony\Contracts\HttpClient\ResponseInterface;
+
+/**
+ * The Lambda function made a network connection to the configured S3 Files access point, but the mount operation timed
+ * out.
+ */
+final class S3FilesMountTimeoutException extends ClientException
+{
+    /**
+     * The exception type.
+     *
+     * @var string|null
+     */
+    private $type;
+
+    public function getType(): ?string
+    {
+        return $this->type;
+    }
+
+    protected function populateResult(ResponseInterface $response): void
+    {
+        $data = $response->toArray(false);
+
+        $this->type = isset($data['Type']) ? (string) $data['Type'] : null;
+    }
+}